@@ -1,17 +1,18 @@
 [build-system]
 requires = ["setuptools>=60", "setuptools-scm>=8.0"]
+build-backend = "setuptools.build_meta"
 
 [tool.setuptools_scm]
-version_file="_version.py"
+version_file = "_version.py"
 
 [tool.black]
 line-length = 95
-target-version = ["py311"]
+target-version = ["py312"]
 extend-exclude = '''^(.*\/)*\b(migrations)\b($|\/.*$)'''
 
 [tool.isort]
 profile = "black"
-py_version = "311"
+py_version = "312"
 skip = [".tox", ".eggs", "migrations"]
 
 [tool.coverage.run]
@@ -35,31 +36,33 @@ exclude_lines = [
 legacy_tox_ini = """
 [tox]
 envlist =
-    py{311}-dj{42,dev},
+    py{312,313}-dj{51,dev},
     lint
+    pre-commit
 
 isolated_build = true
 
 [gh-actions]
 python =
-    3.11: py311, lint
+    3.12: py312, lint, pre-commit
+    3.13: py313
 
 [gh-actions:env]
 DJANGO =
-    4.2: dj42, lint
-    dev: djdev
+    5.1: dj51
+    dev: djdev, lint, pre-commit
 
 [testenv]
 deps =
     -r https://raw.githubusercontent.com/clinicedc/edc/develop/requirements.tests/tox.txt
     -r https://raw.githubusercontent.com/clinicedc/edc/develop/requirements.tests/test_utils.txt
     -r https://raw.githubusercontent.com/clinicedc/edc/develop/requirements.tests/edc.txt
-    -r https://raw.githubusercontent.com/clinicedc/edc/develop/requirements.tests/third_party_dev.txt
-    dj42: Django>=4.2,<5.0
+    dj51: Django>=5.1,<5.2
     djdev: https://github.com/django/django/tarball/main
 
 commands =
-    pip install -U pip coverage[toml]
+    pip install -U pip
+    python --version
     pip --version
     pip freeze
     coverage run -a runtests.py
@@ -68,7 +71,19 @@ commands =
 [testenv:lint]
 deps = -r https://raw.githubusercontent.com/clinicedc/edc/develop/requirements.tests/lint.txt
 commands =
+    python --version
+    pip --version
+    pip freeze
     isort --profile=black --check --diff .
     black --check --diff .
     flake8 .
+
+[testenv:pre-commit]
+deps = pre-commit
+commands =
+    python --version
+    pip --version
+    pip freeze
+    pre-commit autoupdate
+    pre-commit run --all-files
 """
