@@ -12,16 +12,15 @@ keywords = django Edc, forms, markdown, clinicedc, clinical trials
 classifiers=
     Environment :: Web Environment
     Framework :: Django
-    Framework :: Django :: 4.2
+    Framework :: Django :: 5.1
     Intended Audience :: Developers
     Intended Audience :: Science/Research
     Operating System :: OS Independent
-    Programming Language :: Python :: 3.11
+    Programming Language :: Python :: 3.12
     License :: OSI Approved :: GNU General Public License v3 (GPLv3)
 
-
 [options]
-python_requires = >=3.11
+python_requires = >=3.12
 zip_safe = False
 include_package_data = True
 packages = find:
