@@ -1,78 +1,6 @@
 #!/usr/bin/env python
-import logging
-import sys
-from os.path import abspath, dirname, join
 
-import django
-from django.conf import settings
-from django.test.runner import DiscoverRunner
-from edc_test_utils import DefaultTestSettings
-
-app_name = "edc_form_describer"
-base_dir = dirname(abspath(__file__))
-
-DEFAULT_SETTINGS = DefaultTestSettings(
-    calling_file=__file__,
-    APP_NAME=app_name,
-    BASE_DIR=base_dir,
-    ETC_DIR=join(base_dir, app_name, "tests", "etc"),
-    SILENCED_SYSTEM_CHECKS=[
-        "sites.E101",
-        "edc_navbar.E002",
-        "edc_navbar.E003",
-        "edc_consent.E001",
-        "edc_sites.E001",
-        "edc_sites.E002",
-    ],
-    EDC_AUTH_SKIP_SITE_AUTHS=True,
-    EDC_AUTH_SKIP_AUTH_UPDATER=True,
-    INSTALLED_APPS=[
-        "django.contrib.admin",
-        "django.contrib.auth",
-        "django.contrib.contenttypes",
-        "django.contrib.sessions",
-        "django.contrib.messages",
-        "django.contrib.staticfiles",
-        "django.contrib.sites",
-        "django_crypto_fields.apps.AppConfig",
-        "django_revision.apps.AppConfig",
-        "simple_history",
-        "edc_auth.apps.AppConfig",
-        "edc_appointment.apps.AppConfig",
-        "edc_data_manager.apps.AppConfig",
-        "edc_form_runners.apps.AppConfig",
-        "edc_lab.apps.AppConfig",
-        "edc_protocol.apps.AppConfig",
-        "edc_identifier.apps.AppConfig",
-        "edc_device.apps.AppConfig",
-        "edc_metadata.apps.AppConfig",
-        "edc_notification.apps.AppConfig",
-        "edc_offstudy.apps.AppConfig",
-        "edc_randomization.apps.AppConfig",
-        "edc_registration.apps.AppConfig",
-        "edc_sites.apps.AppConfig",
-        "edc_crf.apps.AppConfig",
-        "edc_timepoint.apps.AppConfig",
-        "edc_visit_tracking.apps.AppConfig",
-        "edc_visit_schedule.apps.AppConfig",
-        "edc_action_item.apps.AppConfig",
-        "edc_fieldsets.apps.AppConfig",
-        "edc_form_describer.apps.AppConfig",
-        "edc_appconfig.apps.AppConfig",
-    ],
-    add_dashboard_middleware=True,
-    use_test_urls=True,
-).settings
-
-
-def main():
-    if not settings.configured:
-        settings.configure(**DEFAULT_SETTINGS)
-    django.setup()
-    failures = DiscoverRunner(failfast=True).run_tests([f"{app_name}.tests"])
-    sys.exit(failures)
-
+from edc_test_settings.func_main import func_main2
 
 if __name__ == "__main__":
-    logging.basicConfig()
-    main()
+    func_main2("edc_form_describer.tests.test_settings", "edc_form_describer.tests")
